Abstract out spatial WCS functionality
