+++ spectral_cube/spectral_cube.py
@@ -31,7 +31,7 @@ from .masks import (LazyMask, LazyComparisonMask, BooleanArrayMask, MaskBase,
 from .ytcube import ytCube
 from .lower_dimensional_structures import (Projection, Slice, OneDSpectrum,
                                            LowerDimensionalObject)
-from .base_class import BaseNDClass, SpectralAxisMixinClass, DOPPLER_CONVENTIONS
+from .base_class import BaseNDClass, SpectralAxisMixinClass, DOPPLER_CONVENTIONS, SpatialCoordMixinClass
 from .utils import cached, warn_slow, VarianceWarning, BeamAverageWarning
 
 from distutils.version import LooseVersion
@@ -81,7 +81,7 @@ def aggregation_docstring(func):
 # conventions between WCS and numpy
 np2wcs = {2: 0, 1: 1, 0: 2}
 
-class BaseSpectralCube(BaseNDClass, SpectralAxisMixinClass):
+class BaseSpectralCube(BaseNDClass, SpectralAxisMixinClass, SpatialCoordMixinClass):
 
     def __init__(self, data, wcs, mask=None, meta=None, fill_value=np.nan,
                  header=None, allow_huge_operations=False, wcs_tolerance=0.0):
@@ -257,28 +257,6 @@ class BaseSpectralCube(BaseNDClass, SpectralAxisMixinClass):
         _spectral_max = self.spectral_axis.max()
 
         return _spectral_min, _spectral_max
-
-    @property
-    @cached
-    def world_extrema(self):
-        lat,lon = self.spatial_coordinate_map
-        _lon_min = lon.min()
-        _lon_max = lon.max()
-        _lat_min = lat.min()
-        _lat_max = lat.max()
-
-        return ((_lon_min, _lon_max),
-                (_lat_min, _lat_max))
-
-    @property
-    @cached
-    def longitude_extrema(self):
-        return self.world_extrema[0]
-
-    @property
-    @cached
-    def latitude_extrema(self):
-        return self.world_extrema[1]
 
     def apply_numpy_function(self, function, fill=np.nan,
                              reduce=True, how='auto',
@@ -1468,9 +1446,9 @@ class BaseSpectralCube(BaseNDClass, SpectralAxisMixinClass):
         """
         return spectral_axis.determine_vconv_from_ctype(self.wcs.wcs.ctype[self.wcs.wcs.spec])
 
-    @property
-    def spatial_coordinate_map(self):
-        return self.world[0, :, :][1:]
+    # @property
+    # def spatial_coordinate_map(self):
+    #     return self.world[0, :, :][1:]
 
     def closest_spectral_channel(self, value):
         """
@@ -1801,69 +1779,6 @@ class BaseSpectralCube(BaseNDClass, SpectralAxisMixinClass):
         # by using ceil / floor above, we potentially introduced a NaN buffer
         # that we can now crop out
         return masked_subcube.minimal_subcube(spatial_only=True)
-
-
-    def world_spines(self):
-        """
-        Returns a list of 1D arrays, for the world coordinates
-        along each pixel axis.
-
-        Raises error if this operation is ill-posed (e.g. rotated world coordinates,
-        strong distortions)
-
-        This method is not currently implemented. Use :meth:`world` instead.
-        """
-        raise NotImplementedError()
-
-    @cube_utils.slice_syntax
-    def world(self, view):
-        """
-        Return a list of the world coordinates in a cube (or a view of it).
-
-        Cube.world is called with *bracket notation*, like a NumPy array::
-            c.world[0:3, :, :]
-
-        Returns
-        -------
-        [v, y, x] : list of NumPy arrays
-            The 3 world coordinates at each pixel in the view.
-
-        Examples
-        --------
-        Extract the first 3 velocity channels of the cube:
-        >>> v, y, x = c.world[0:3]
-
-        Extract all the world coordinates
-        >>> v, y, x = c.world[:, :, :]
-
-        Extract every other pixel along all axes
-        >>> v, y, x = c.world[::2, ::2, ::2]
-        """
-
-        # note: view is a tuple of view
-
-        # the next 3 lines are equivalent to (but more efficient than)
-        # inds = np.indices(self._data.shape)
-        # inds = [i[view] for i in inds]
-        inds = np.ogrid[[slice(0, s) for s in self._data.shape]]
-        inds = np.broadcast_arrays(*inds)
-        inds = [i[view] for i in inds[::-1]]  # numpy -> wcs order
-
-        shp = inds[0].shape
-        inds = np.column_stack([i.ravel() for i in inds])
-        world = self._wcs.all_pix2world(inds, 0).T
-
-        world = [w.reshape(shp) for w in world]  # 1D->3D
-
-        # apply units
-        world = [w * u.Unit(self._wcs.wcs.cunit[i])
-                 for i, w in enumerate(world)]
-
-        # convert spectral unit if needed
-        if self._spectral_unit is not None:
-            world[2] = world[2].to(self._spectral_unit)
-
-        return world[::-1]  # reverse WCS -> numpy order
 
     def _val_to_own_unit(self, value, operation='compare', tofrom='to',
                          keepunit=False):
